+++ csharp-advanced-topics/ActionFuncDelegatesInCsharp/ActionFuncDelegatesInCsharp.sln
@@ -1,0 +1,37 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.0.31903.59
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ActionDelegateInCsharp", "ActionDelegateInCsharp\ActionDelegateInCsharp.csproj", "{A345CAF6-1F5F-49C1-AF32-E4967FB78059}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FuncDelegateInCsharp", "FuncDelegateInCsharp\FuncDelegateInCsharp.csproj", "{B4C61D09-FBA2-468C-988E-35C3B209E3DE}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Tests", "Tests\Tests.csproj", "{100621AE-4EB8-401D-A83B-8FC27190B1BE}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{A345CAF6-1F5F-49C1-AF32-E4967FB78059}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{A345CAF6-1F5F-49C1-AF32-E4967FB78059}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{A345CAF6-1F5F-49C1-AF32-E4967FB78059}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{A345CAF6-1F5F-49C1-AF32-E4967FB78059}.Release|Any CPU.Build.0 = Release|Any CPU
+		{B4C61D09-FBA2-468C-988E-35C3B209E3DE}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B4C61D09-FBA2-468C-988E-35C3B209E3DE}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B4C61D09-FBA2-468C-988E-35C3B209E3DE}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B4C61D09-FBA2-468C-988E-35C3B209E3DE}.Release|Any CPU.Build.0 = Release|Any CPU
+		{100621AE-4EB8-401D-A83B-8FC27190B1BE}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{100621AE-4EB8-401D-A83B-8FC27190B1BE}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{100621AE-4EB8-401D-A83B-8FC27190B1BE}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{100621AE-4EB8-401D-A83B-8FC27190B1BE}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {031956B3-3F89-4C2A-AD12-E345D8814317}
+	EndGlobalSection
+EndGlobal
