+++ csharp-advanced-topics/ActionFuncDelegatesInCsharp/Tests/Usings.cs
@@ -1,0 +1,3 @@
+global using Xunit;
+global using ActionDelegateInCsharp;
+global using FuncDelegateInCsharp;
