--- csharp-intermediate-topics/ActionFuncDelegatesInCsharp/ActionFuncDelegatesInCsharp.sln
@@ -1,31 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.0.31903.59
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ActionFuncDelegatesInCsharp", "ActionFuncDelegatesInCsharp\ActionFuncDelegatesInCsharp.csproj", "{3785E736-A3EC-4004-9732-16B96385DEF0}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ActionFuncDelegatesTest", "ActionFuncDelegatesTest\ActionFuncDelegatesTest.csproj", "{7D540311-0C04-4A60-9956-8C86AF401071}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{3785E736-A3EC-4004-9732-16B96385DEF0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{3785E736-A3EC-4004-9732-16B96385DEF0}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{3785E736-A3EC-4004-9732-16B96385DEF0}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{3785E736-A3EC-4004-9732-16B96385DEF0}.Release|Any CPU.Build.0 = Release|Any CPU
-		{7D540311-0C04-4A60-9956-8C86AF401071}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{7D540311-0C04-4A60-9956-8C86AF401071}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{7D540311-0C04-4A60-9956-8C86AF401071}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{7D540311-0C04-4A60-9956-8C86AF401071}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {603F34B1-B87F-4136-A309-E1A15DC708ED}
-	EndGlobalSection
-EndGlobal
