--- csharp-intermediate-topics/ActionFuncDelegatesInCsharp/ActionFuncDelegatesInCsharp/Program.cs
@@ -1,2 +1,0 @@
-﻿// See https://aka.ms/new-console-template for more information
-Console.WriteLine("Hello, World!");
