--- csharp-intermediate-topics/ActionFuncDelegatesInCsharp/ActionFuncDelegatesTest/Usings.cs
@@ -1,2 +1,0 @@
-global using Xunit;
-global using ActionFuncDelegatesInCsharp;
